+++ src/main/java/com/gildedgames/util/ui/util/Font.java
@@ -5,10 +5,10 @@ import java.util.List;
 public interface Font
 {
 	
-	double getWidth(String text);
+	float getWidth(String text);
 
-	double getHeight(String text);
+	float getHeight(String text);
 
-	List<String> splitStringsIntoArea(String text, double width);
+	List<String> splitStringsIntoArea(String text, float width);
 
 }
